refactor(server-stats): change RAM and CPU usage fields to float for precision
This commit updates the data type of RAM and CPU usage fields from `long` to `float` to improve precision in representing usage statistics. Additionally, the `updateStats` method in `ServerService` is refactored to calculate CPU percentage and convert memory usage to MB for better readability and accuracy. The `serverKillFlags` field is also made private to enforce encapsulation.

+++ src/main/java/mindustrytool/servermanager/messages/response/StatsMessageResponse.java
@@ -9,9 +9,9 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 @Data
 public class StatsMessageResponse {
-    public long ramUsage = 0;
-    public long totalRam = 0;
-    public long cpuUsage = 0l;
+    public float ramUsage = 0;
+    public float totalRam = 0;
+    public float cpuUsage = 0l;
     public long players = 0;
     public String mapName = "";
     public byte[] mapData;
